Added interfaces folder and one class per file
