+++ AElf.Kernel/Storages/Change.cs
@@ -1,0 +1,8 @@
+﻿namespace AElf.Kernel.Storages
+{
+    public struct Change
+    {
+        public Path Before { get; set; }
+        public Path After { get; set; }
+    }
+}
