+++ AElf.Kernel/Storages/Interfaces/IChainBlockRelationStore.cs
@@ -1,0 +1,10 @@
+﻿using System.Threading.Tasks;
+
+namespace AElf.Kernel.Storages
+{
+    public interface IChainBlockRelationStore
+    {
+        Task InsertAsync(Chain chain, Block block);
+        Task<Hash> GetAsync(Hash chainHash, long height);
+    }
+}
