+++ AElf.Kernel/Storages/Interfaces/IChainStore.cs
@@ -1,0 +1,11 @@
+﻿using System.Threading.Tasks;
+
+namespace AElf.Kernel.Storages
+{
+    public interface IChainStore
+    {
+        Task<Chain> GetAsync(Hash id);
+        Task<Chain> UpdateAsync(Chain chain);
+        Task<Chain> InsertAsync(Chain chain);
+    }
+}
