+++ AElf.Kernel/Storages/Interfaces/IChangesStore.cs
@@ -1,0 +1,9 @@
+﻿using System.Threading.Tasks;
+
+namespace AElf.Kernel.Storages
+{
+    public interface IChangesStore
+    {
+        Task InsertAsync(Hash path, Change before);
+    }
+}
