+++ AElf.Kernel/Storages/Interfaces/IPointerStore.cs
@@ -1,0 +1,10 @@
+﻿using System.Threading.Tasks;
+
+namespace AElf.Kernel.Storages
+{
+    public interface IPointerStore
+    {
+        Task InsertAsync(Hash path, Hash pointer);
+        Task<Hash> GetAsync(Hash path);
+    }
+}
