+++ AElf.Kernel/Storages/Interfaces/ITransactionStore.cs
@@ -1,0 +1,10 @@
+﻿using System.Threading.Tasks;
+
+namespace AElf.Kernel.Storages
+{
+    public interface ITransactionStore
+    {
+        Task InsertAsync(Transaction tx);
+        Task<Transaction> GetAsync(Hash hash);
+    }
+}
