+++ AElf.Kernel/Storages/KeyValueDatabase.cs
@@ -1,16 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using AElf.Kernel.Extensions;
 
 namespace AElf.Kernel.Storages
 {
-    public interface IKeyValueDatabase
-    {
-        Task<object> GetAsync(Hash key,Type type);
-        Task SetAsync(Hash key, object bytes);
-    }
-
     public class KeyValueDatabase : IKeyValueDatabase
     {
         private readonly Dictionary<Hash, object> _dictionary = new Dictionary<Hash, object>();
